a very bad design
