--- src/index.py
@@ -1,3 +1,0 @@
-adlkfjskl
-
-asdf
